==> scripts/04_pancreatic_marker_characterization.Rmd <==

## Pancreatic Cancer Database

```{r}
files <- list.files("data/Other/pcd_download")
files <- files[-2471] # empty file

pcd <- 
  map_df(files, function(file) {
    
    result <<- xmlToDataFrame(paste0("data/Other/pcd_download/", file))
    
    protein_name <- result[2,1]
    
    protein_res <- 
      result |> 
      select(any_of(c("cancer_type", "Experiment_Type", "Expression_Level", "Fold_Change", "PubMed"))) |> 
      mutate(Protein = protein_name) |> 
      relocate(Protein)
    
  }) 

pcd |> 
  distinct(Protein) |> 
  nrow()

pcd_all <- 
  pcd |> 
  filter(Expression_Level %in% c("Increased", "Decreased"))

pcd_proteins <- 
  pcd |> 
  filter(Expression_Level %in% c("Increased", "Decreased")) |> 
  distinct(Protein) |> 
  pull()

pcd_all |> 
  write_excel_csv(savepath("pcd.csv"))
```



## HPA databases

```{r}

hpa_info <- read_tsv("data/hpa/proteinatlas.tsv")

columns_keep <- c("Gene", "Ensembl", "Gene description", "Protein class", "Biological process",
                  "Molecular function", "Disease involvement", "RNA tissue specificity", "RNA tissue specific nTPM",
                  "Pathology prognostics - Pancreatic cancer", "Subcellular main location",
                  "Tissue expression cluster", "Brain expression cluster",
                  "Blood concentration - Conc. blood IM [pg/L]", "Secretome location",
                  "Secretome function")  

hpa_info_short <- 
  hpa_info |> 
  select(all_of(columns_keep))

hpa_rna <-  read_tsv("data/hpa/rna_tissue_consensus.tsv")
hpa_protein <-  read_tsv("data/hpa/normal_tissue.tsv")
hpa_pathology <- read_tsv("data/hpa/pathology.tsv") 
hpa_tcga <- read_tsv("data/hpa/rna_cancer_sample_109.tsv")

hpa_protein_class <- read_tsv("data/hpa/proteinclass_103.tsv")
hpa_enhanced_tissues <- read_tsv("data/hpa/consensus_all_category_103.tsv")
GO_terms <- read_tsv("data/hpa/Ensembl103 GO terms.txt")
reactome <- 
  read_tsv("data/hpa/Ensembl2Reactome_All_Levels_20210903.txt") |> 
  filter(species == "Homo sapiens")

```



# Protein summary

## Prepare ranking

```{r}
importances <- 
  readRDS("data/processed/ML_res/Importances_combined.rds")

de <- 
  readRDS("data/processed/DE/DE_combined.rds")

all_rank <- 
  de |> 
  left_join(importances, by = c("Assay" = "Variable")) |> 
  mutate(rank_de_healthy = rank(-abs(logFC_healthy)),
         rank_de_cancers = rank(-abs(logFC_cancers)),
         rank_rf = rank(-Importance_rf),
         rank_lasso = rank(-Importance_lasso)) |> 
  mutate(sum_rank = rank_de_healthy + rank_de_cancers + rank_rf + rank_lasso) |> 
  arrange(sum_rank)

all_rank |> 
  mutate(pcd = ifelse(Assay %in% pcd_proteins, "Yes", "No")) |> 
  relocate(pcd, 1) |> 
  View()

all_rank |> 
  distinct(Assay) |> 
  left_join(data_NPX |> 
              distinct(Assay, Panel), by = "Assay")
#236 proteins have both info
# 531 down -> 46
# 2,792 up -> 365

# Highlight in volcano plot

top_30 <- 
  all_rank |> 
  head(30) |> 
  pull(Assay)

top_200 <- 
  all_rank |> 
  head(100) |> 
  pull(Assay)

saveRDS(top_30, "data/processed/protein_lists/top30_pancreatic.rds")
```


## Comparison DE healthy/cancers

```{r}
all_res_DA <- 
  de_healthy |> 
  select(Assay, logFC, adj.P.Val, sig) |> 
  mutate(Type = "DA_Healthy") |> 
  bind_rows(de_cancers |> 
  select(Assay, logFC, adj.P.Val, sig) |> 
  mutate(Type = "DA_Cancers")) |> 
  left_join(hpa_info_short |> 
  select(Gene, secreted =`Secretome location`), by = c("Assay" = "Gene"))


alluvial_dat_healthy <- 
  all_res_DA |> 
  filter(Type == "DA_Healthy") |> 
  group_by(sig, secreted) |> 
  summarise(n = n_distinct(Assay)) |> 
  mutate(secreted = ifelse(is.na(secreted), "Unknown", secreted))

alluvial_dat_cancer <- 
  all_res_DA |> 
  filter(Type == "DA_Cancers") |> 
  group_by(sig, secreted) |> 
  summarise(n = n_distinct(Assay)) |> 
  mutate(secreted = ifelse(is.na(secreted), "Unknown", secreted))


library(ggalluvial)


unique(alluvial_dat$secreted)


summary_data <- 
  alluvial_dat_healthy %>%
  filter(sig != "not significant") |>
  mutate(sig = factor(sig, levels = c("significant up", "significant down")),
         secreted = factor(secreted, levels = names(pal_secreted))) |>
  ungroup() |> 
  arrange(desc(sig), desc(secreted)) |> 
 # group_by(sig, secreted) %>%
  mutate(total_proteins = cumsum(n))

alluvial_dat_healthy |> 
  filter(sig != "not significant") |> #pull(n) |> sum()
  mutate(sig = factor(sig, levels = c("significant up", "significant down")),
         secreted = factor(secreted, levels = names(pal_secreted))) |> 
  ggplot(aes(axis1 = sig, axis2 = secreted, y = n, fill = secreted)) +
  geom_alluvium(aes(fill = secreted)) +
  geom_stratum() +
  geom_text(stat = "stratum", aes(label = after_stat(stratum))) +
  geom_text(data = summary_data, 
            aes(label = n, 
                x = "sig", y = total_proteins, color = secreted), 
            hjust = 1, size = 4, fontface = "bold") +
  scale_x_discrete(limits = c("sig", "secreted")) +
  scale_y_continuous(limits = c(0,208)) +
  scale_fill_manual(values = pal_secreted, breaks = names(pal_secreted)) +
  scale_color_manual(values = pal_secreted, breaks = names(pal_secreted)) +
  theme_void()

ggsave(savepath("alluvial_healthy.pdf"), h = 5, w = 7)

summary_data <- 
  alluvial_dat_cancer %>%
  filter(sig != "not significant") |>
  mutate(sig = factor(sig, levels = c("significant up", "significant down")),
         secreted = factor(secreted, levels = names(pal_secreted))) |>
  ungroup() |> 
  arrange(desc(sig), desc(secreted)) |> 
 # group_by(sig, secreted) %>%
  mutate(total_proteins = cumsum(n))

alluvial_dat_cancer |> 
  filter(sig != "not significant") |> #pull(n) |> sum()
  mutate(sig = factor(sig, levels = c("significant up", "significant down")),
         secreted = factor(secreted, levels = names(pal_secreted))) |> 
  arrange(sig, secreted) |> 
  ggplot(aes(axis1 = sig, axis2 = secreted, y = n, fill = secreted)) +
  geom_alluvium() +
  geom_stratum() +
  geom_text(stat = "stratum", aes(label = after_stat(stratum))) +
  geom_text(data = summary_data, 
            aes(label = n, 
                x = "sig", y = total_proteins, color = secreted), 
            hjust = 1, size = 4, fontface = "bold") +
  scale_x_discrete(limits = c("sig", "secreted")) +
  scale_y_continuous(limits = c(0, 93)) +
  scale_fill_manual(values = pal_secreted, breaks = names(pal_secreted)) +
  scale_color_manual(values = pal_secreted, breaks = names(pal_secreted)) +
  theme_void()

ggsave(savepath("alluvial_cancer.pdf"), h = 5, w = 7)


# Euler diagrams
library(eulerr)

up_healthy <- 
  de_healthy |> 
  filter(sig == "significant up") |> 
  pull(Assay)

up_cancers <- 
  de_cancers |> 
  filter(sig == "significant up") |> 
  pull(Assay)

y <- list("Healthy" = up_healthy, 
          "Cancers" = up_cancers)
plot(euler(y, shape = "ellipse"), quantities = TRUE) |> as.ggplot()

ggsave(savepath("overlap_upregulated.pdf"), h = 3, w = 3)

down_healthy <- 
  de_healthy |> 
  filter(sig == "significant down") |> 
  pull(Assay)
 
down_cancers <- 
  de_cancers |> 
  filter(sig == "significant down") |> 
  pull(Assay)

y <- list("Healthy" = down_healthy, 
          "Cancers" = down_cancers)
plot(euler(y, shape = "ellipse"), quantities = TRUE) |> as.ggplot()

ggsave(savepath("overlap_downregulated.pdf"), h = 3, w = 3)


```

## DE summary

```{r}

some_prots <- c("LTA4H", "KLK13", "GDF15", "SMOC1", "ITGAV", "ADAMTS8", "AGRP", "SPP1", "LRRN1", "NME3")
  
all_res <- 
  de_healthy |> 
  select(Assay, logFC, adj.P.Val) |> 
  mutate(Type = "DA_Healthy") |> 
  bind_rows(de_cancers |> 
  select(Assay, logFC, adj.P.Val) |> 
  mutate(Type = "DA_Cancers")) |> 
  bind_rows(de_cptac |> 
  select(Assay, logFC, adj.P.Val) |> 
  mutate(Type = "CPTAC")) |> 
  bind_rows(de_results_ukb_1 |> 
  select(Assay, logFC, adj.P.Val) |> 
  mutate(Type = "UKB_1_year")) |> 
  bind_rows(de_results_ukb_3 |> 
  select(Assay, logFC, adj.P.Val) |> 
  mutate(Type = "UKB_3_year")) |> 
  bind_rows(de_results_ukb_5 |> 
  select(Assay, logFC, adj.P.Val) |> 
  mutate(Type = "UKB_5_year"))



all_res |> 
  filter(Assay %in% some_prots) |> 
  mutate(Type = factor(Type, levels = unique(all_res$Type))) |> 
  ggplot(aes(Type, Assay, fill = logFC#, size = -log10(adj.P.Val))
         )) +
  geom_point(color = "black", shape = 21, size = 5) +
  scale_fill_distiller(palette = "RdBu") +
#  scale_color_brewer("RdBu") +
#  scale_color_manual(colorRampPalette(brewer.pal(8, "RdBu"))(100)) +
  theme_hpa(angled = T)









```

## Table 


```{r}

# hpa_info_short |> 
#   filter(Gene %in% important_proteins_da$term) |> View()
# 
# important_proteins_da |> 
#   select(-penalty) |> 
#   rename(Protein = term, `Model coefficient` = estimate) |> 
#   left_join(de_combined |> 
#               select(Protein = Assay, Significance = sig_healthy, logFC = logFC_healthy), by = "Protein") |> 
#   left_join(, by = )

hpa_info_short |> 
  select(Protein = Gene, 
         `Gene description`, 
         `Protein class`, 
         `RNA tissue specificity`, 
         `Secretome location`) |> 
  filter(Protein %in% all_rank$Assay) |> 
  mutate(Protein = factor(Protein, levels = all_rank$Assay)) |> 
  arrange(Protein) |> 
  left_join(all_rank |> 
              select(-rank_de_cancers, -rank_de_healthy, -rank_rf, -rank_lasso, -sum_rank, -Significance, -Type) |> 
              rename(Protein = Assay), by = "Protein") |> 
  write_csv(savepath("all_rank_summary.csv"))




#|> 
  mutate_if(is.numeric, round, 2) |> 
  gt() |> 
  data_color(
    columns = `Model coefficient`,
    method = "numeric",
    palette = "ggsci::red_material"
  ) |> 
 
  # data_color(columns = vars(logFC_healthy:logFC_cancers), 
  #            colors = pal_de)
  # data_color(columns = vars(logFC_healthy:logFC_cancers),
  #   direction = "row",
  #   palette = "PuOr",
  #   na_color = "white") |>
  opt_all_caps() |>
  opt_vertical_padding(scale = 0.75) |>
  opt_horizontal_padding(scale = 3) |>
  tab_options(
    table_body.hlines.style = "none",
    column_labels.border.top.color = "black",
    column_labels.border.bottom.color = "black",
    table_body.border.bottom.color = "black"
  ) 
                         
all_rank |> 
  select(-rank_de_cancers, -rank_de_healthy, -rank_rf, -rank_lasso, -sum_rank, -Significance, -Type) |> 
  #mutate_if(is.numeric, round, 2) |> 
  gt() |> 
  # data_color(
  #  # columns = `Model coefficient`,
  #   method = "numeric",
  #   palette = "ggsci::red_material"
  # ) |> 
 
  # data_color(columns = vars(logFC_healthy:logFC_cancers), 
  #            colors = pal_de)
  # data_color(columns = vars(logFC_healthy:logFC_cancers),
  #   direction = "row",
  #   palette = "PuOr",
  #   na_color = "white") |>
  opt_all_caps() |>
  opt_vertical_padding(scale = 0.75) |>
  opt_horizontal_padding(scale = 3) |>
  tab_options(
    table_body.hlines.style = "none",
    column_labels.border.top.color = "black",
    column_labels.border.bottom.color = "black",
    table_body.border.bottom.color = "black"
  ) 

##                          

important_proteins |> 
  mutate(Sign = ifelse(estimate > 0, "Positive", "Negative"),
         estimate = abs(estimate)) |> 
  ggplot(aes(fct_reorder(term, abs(estimate)), estimate, fill = Sign)) +
  geom_col(alpha = 0.7, position = "dodge") +
  facet_wrap(~Sign, scales = "free") +
  scale_fill_manual(values = c("Negative" = "#8267AB",
                               "Positive" = "#ADDD8E")) +
  coord_flip() +
  labs(x = "Protein", y = "Coefficient") +
  theme_main

ggsave(savepath("important_prots.pdf"), h = 4, w = 7)

ann <- 
  cancer_data_wide |> 
  filter(Disease %in% c("PAN", "CAD healthy")) |> 
  mutate(Disease = case_when(Disease == "PAN" ~ "Pancreatic cancer",
                             Disease == "CAD healthy" ~ "Healthy")) |> 
  select(DAid, Disease) |> 
  column_to_rownames("DAid")

ann_cols <- 
  important_proteins |> 
  mutate(Sign = ifelse(estimate > 0, "Positive", "Negative"),
         estimate = abs(estimate)) |> 
  select(term, Sign #, Estimate = estimate
         ) |> 
  column_to_rownames("term")

library(RColorBrewer)
rd_bu_continuous <- rev(colorRampPalette(brewer.pal(11, "RdBu"))(100))

cancer_data_wide |> 
  filter(Disease %in% c("PAN", "CAD healthy")) |> 
  select(DAid, important_proteins$term) |> 
  column_to_rownames("DAid") |> 
  scale() |> 
  pheatmap(show_rownames = F,
           annotation_row = ann,
           annotation_col = ann_cols,
           color = rd_bu_continuous,
           annotation_colors = list("Disease" = pal_group[-3],
                                    "Sign" = c("Negative" = "#8267AB", "Positive" = "#ADDD8E")#,
                                    #"Estimate" = colorRampPalette(c("white", "#8267AB"))(100))
           )) |> 
  as.ggplot()

ggsave(savepath("ML_heatmap.pdf"), h = 4, w = 8)


# Boxplots

top_up <- c("LTA4H", "AGRP", "CSF1", "TFRC")
top_down <-  c("ADAMTS8", "KLK13", "RGMA", "HPGDS")

long_levels <- 
  cancers_mapping |> 
  mutate(Cancer = ifelse(Cancer_code %in% c("AML", "CLL"), Cancer_code, Cancer),
         Cancer = ifelse(Cancer_code %in% c("LYMPH"), "DLBCL", Cancer),
         Cancer_code = factor(Cancer_code, levels = all_cancers_levels)) |> 
  arrange(Cancer_code) |> 
  pull(Cancer)

upregulated_prots <- 
  cancer_data_long |> 
  filter(Assay %in% top_up,
         Disease %in% all_levels) |> 
  rename(GROUP = Disease) |> 
  mutate(Disease = case_when(GROUP == "PAN" ~ "Pancreatic cancer",
                             GROUP == "CAD healthy" ~ "Healthy",
                             T ~ "Other cancers")) |> 
  left_join(cancers_mapping |> 
              mutate(Cancer = ifelse(Cancer_code %in% c("AML", "CLL"), Cancer_code, Cancer),
                     Cancer = ifelse(Cancer_code %in% c("LYMPH"), "DLBCL", Cancer)), 
            by = c("GROUP" = "Cancer_code")) |> 
  mutate(Cancer = ifelse(GROUP == "CAD healthy", "Healthy", Cancer),
         Assay = factor(Assay, levels = top_up),
         Cancer = factor(Cancer, levels = c(long_levels, "Healthy"))) |> 
  select(DAid, Cancer, Disease, Assay, NPX) |> 
  ggplot(aes(Cancer, NPX, color = Disease)) +
  geom_quasirandom(alpha = 0.4, show.legend = F) +
  geom_boxplot(fill = NA, color = "grey60", outlier.color = NA) +
  facet_wrap(~ Assay, ncol = 2, scales = "free_y") +  
  scale_color_manual(values = pal_group) +
  xlab("") +
  theme_simple


downregulated_prots <- 
  cancer_data_long |> 
  filter(Assay %in% top_down,
         Disease %in% all_levels) |> 
  rename(GROUP = Disease) |> 
  mutate(Disease = case_when(GROUP == "PAN" ~ "Pancreatic cancer",
                             GROUP == "CAD healthy" ~ "Healthy",
                             T ~ "Other cancers")) |> 
  left_join(cancers_mapping |> 
              mutate(Cancer = ifelse(Cancer_code %in% c("AML", "CLL"), Cancer_code, Cancer),
                     Cancer = ifelse(Cancer_code %in% c("LYMPH"), "DLBCL", Cancer)), 
            by = c("GROUP" = "Cancer_code")) |> 
  mutate(Cancer = ifelse(GROUP == "CAD healthy", "Healthy", Cancer),
         Assay = factor(Assay, levels = top_down),
         Cancer = factor(Cancer, levels = c(long_levels, "Healthy"))) |> 
  select(DAid, Cancer, Disease, Assay, NPX) |> 
  ggplot(aes(Cancer, NPX, color = Disease)) +
  geom_quasirandom(alpha = 0.4, show.legend = F) +
  geom_boxplot(fill = NA, color = "grey60", outlier.color = NA) +
  facet_wrap(~ Assay, ncol = 2, scales = "free_y") +  
  scale_color_manual(values = pal_group) +
  xlab("") +
  theme_simple

downregulated_prots + upregulated_prots 

ggsave(savepath("boxplots_ml.pdf"), h = 7, w = 16)


lasso_important_proteins <- 
  final_glmnet_fit  |> 
  extract_fit_parsnip() %>%
  tidy() |> 
  filter(term != "(Intercept)") |> 
  arrange(-abs(estimate)) 

dat <- 
  lasso_important_proteins |>
  select(Assay = term, estimate) |>
  left_join(de_results |>
              select(Assay, logFC, adj.P.Val), by = "Assay") 

dat |>
  ggplot(aes(logFC, estimate, size = -log10(adj.P.Val),
             color = abs(estimate))) +
  geom_point(alpha = 0.8) +
  scale_color_viridis() +
  geom_text_repel(data = dat |> filter(abs(estimate) > 0),
                           aes(label = Assay), show.legend = F)

ggsave(savepath("current_de.png"))

```

```{r}

saveRDS(important_proteins, "data/processed/pancreatic/important_proteins_lasso.rds")

# Figure heatmap
importance <- 
  important_proteins |> 
  mutate(Sign = ifelse(estimate > 0, "Positive", "Negative"),
         estimate = abs(estimate)) |>
  arrange(Sign, -estimate) |> 
  mutate(term = factor(term, levels = (term))) |> 
  ggplot(aes(term, estimate, fill = Sign)) +
  geom_col(alpha = 0.7, position = "dodge", show.legend = F) +
  scale_fill_manual(values = c("Negative" = "#8267AB",
                               "Positive" = "#ADDD8E")) +
  labs(x = "Protein", y = "Coefficient") +
  theme_main +
  theme(axis.title.x = element_blank(),
        axis.text.x = element_blank(),
        axis.ticks.x = element_blank(),
        axis.line.x = element_blank())

ggsave(savepath("importance_2.pdf"), h = 2, w= 8)   
theme_main <- 
  theme(panel.grid.major = element_blank(), 
                     panel.grid.minor = element_blank(),
                     panel.spacing = unit(0.2, "lines"), 
                     panel.background=element_rect(fill="white"),
         panel.border = element_blank(),
         plot.title = element_text(face = "bold",
                                              size = rel(1), hjust = 0.5),
            plot.subtitle=element_text(face = "bold",hjust = 0.5, size=rel(1),vjust=1),
         axis.title = element_text(face = "bold",size = rel(1)),
                     axis.ticks = element_line(),
                     axis.ticks.length = unit(.25, "cm"),
                     axis.line = element_line(colour = "black", linewidth = rel(0.5)),
                     axis.text = element_text(size = rel(1), color = 'black'),
                     legend.key = element_blank(),
                     legend.position = "right",
                     legend.text = element_text(size=rel(0.8)),
                     legend.key.size= unit(0.7, "cm"),
                     legend.title = element_text(size=rel(1)),
                   plot.margin=unit(c(10,5,5,5),"mm"),
                     strip.background=element_rect(colour="grey90",fill="grey90"),
                    strip.text = element_text(face="bold")
) 

ggsave(savepath("test.png"))
ann <- 
  cancer_data_wide |> 
  filter(Disease %in% c("PAN", "CAD healthy")) |> 
  mutate(Disease = case_when(Disease == "PAN" ~ "Pancreatic cancer",
                             Disease == "CAD healthy" ~ "Healthy")) |> 
  select(DAid, Disease) |> 
  column_to_rownames("DAid")

library(RColorBrewer)
ann_cols <- 
  important_proteins |> 
  mutate(Sign = ifelse(estimate > 0, "Positive", "Negative"),
         estimate = abs(estimate)) |>
  arrange(Sign, -estimate) |> 
  select(term, Sign) |> 
  column_to_rownames("term")

rd_bu_continuous <- rev(colorRampPalette(brewer.pal(11, "RdBu"))(100))

prot_heatmap <- 
  cancer_data_wide |>
  select(DAid, Disease, rownames(ann_cols)) |> 
  filter(Disease %in% c("PAN", "CAD healthy")) |> 
  select(-Disease) |> 
  column_to_rownames("DAid") |> 
  scale() |> 
  pheatmap(show_rownames = F,
           cluster_cols = F,
           annotation_row = ann,
           annotation_col = ann_cols,
           color = rd_bu_continuous,
           annotation_colors = list("Disease" = pal_group[-3],
                                    "Sign" = c("Negative" = "#8267AB", "Positive" = "#ADDD8E")#,
                                    #"Estimate" = colorRampPalette(c("white", "#8267AB"))(100))
           )) |> 
  as.ggplot()




ggsave(savepath("ML_heatmap.pdf"), h = 4, w = 8)

importance / prot_heatmap


# Figure boxplots
top_up <- 
  important_proteins |> 
  top_n(4, estimate) |> 
  pull(term)
top_down <-  
  important_proteins |> 
  top_n(4, -estimate) |> 
  pull(term)


cancer_data_long |> 
  filter(Assay %in% c(top_up, top_down),
         Disease %in% all_levels) |> 
  rename(GROUP = Disease) |> 
  mutate(Disease = case_when(GROUP == "PAN" ~ "Pancreatic cancer",
                             GROUP == "CAD healthy" ~ "Healthy",
                             T ~ "Other cancers")) |> 
  left_join(cancers_mapping |> 
              mutate(Cancer = ifelse(Cancer_code %in% c("AML", "CLL"), Cancer_code, Cancer),
                     Cancer = ifelse(Cancer_code %in% c("LYMPH"), "DLBCL", Cancer)), 
            by = c("GROUP" = "Cancer_code")) |> 
  mutate(Cancer = ifelse(GROUP == "CAD healthy", "Healthy", Cancer),
         Assay = factor(Assay, levels = c(top_down, top_up)),
         Cancer = factor(Cancer, levels = c(long_levels, "Healthy"))) |> 
  select(DAid, Cancer, Disease, Assay, NPX) |> 
  ggplot(aes(Cancer, NPX, color = Disease)) +
  geom_quasirandom(alpha = 0.4, show.legend = F) +
  geom_boxplot(fill = NA, color = "grey60", outlier.color = NA) +
  facet_wrap(~ Assay, ncol = 8, scales = "free_x") +  
  scale_color_manual(values = pal_group) +
  coord_flip() +
  xlab("") +
  theme_simple

ggsave(savepath("boxplots_ml.pdf"), h = 4, w = 16)


```

## Figure

```{r}
all_info <- 
  hpa_info_short |> 
  select(Protein = Gene, 
         `Gene description`, 
         `Protein class`, 
         `RNA tissue specificity`,
         `RNA tissue specific nTPM`,
         `Secretome location`) |> 
  filter(Protein %in% all_rank$Assay) |> 
  mutate(Protein = factor(Protein, levels = all_rank$Assay)) |> 
  arrange(Protein) |> 
  left_join(all_rank |> 
              select(-rank_de_cancers, -rank_de_healthy, -rank_rf, -rank_lasso, -sum_rank, -Significance, -Type) |> 
              rename(Protein = Assay), by = "Protein") |> 
  mutate(Cancer_related = ifelse(grepl("Cancer-related", `Protein class`), "Yes", "No"),
         Enriched_pancreas = ifelse(`RNA tissue specificity` %in% c("Tissue enhanced", "Tissue enriched") & grepl("pancreas", `RNA tissue specific nTPM`), "Yes", "No")) |> 
  select(-`Protein class`) |> 
  mutate(Pancreatic_cancer_databse = ifelse(Protein %in% pcd_proteins, "Yes", "No"))

p0 <- 
  all_info |> 
  head(30) |> 
  mutate(Protein = factor(Protein, levels = rev(Protein)),
         `Gene description` = factor(`Gene description`, levels = rev(`Gene description`))) |> 
  select(Protein, `Gene description`) |> 
  arrange(Protein) |> 
  ggplot(aes(x = "", y = `Gene description`)) +
  geom_text(aes(label = Protein), hjust = 0.5, fontface = "bold", size = 3) +
  theme_void() +
  theme(axis.text.y = element_text(hjust = 1, size = 8))


p1 <- 
  all_info |> 
  head(30) |> 
  mutate(Protein = factor(Protein, levels = rev(Protein)),
         Secreted_blood = ifelse(`Secretome location` == "Secreted to blood", "Yes", "No")) |> 
  select(Protein, `Gene description`, PC_database = Pancreatic_cancer_databse, Cancer_related,Secreted_blood,
         Enriched_pancreas) |>
  pivot_longer(-c(1:2), names_to = "Type", values_to = "Status") |> 
  mutate(Type = factor(Type, levels = c("PC_database",
                                        "Cancer_related",
                                        "Secreted_blood",
                                        "Enriched_pancreas"))) |> 
  arrange(Protein) |> 
  ggplot(aes(Type, Protein, fill = Status)) +
  geom_tile(color = "white") +
  scale_fill_manual(values = c("grey", "red")) +
  coord_equal() +
  theme_hpa(angled = T) +
  theme(axis.ticks.y = element_blank(),
        axis.line.y = element_blank(),
        axis.text.y = element_blank(),
        axis.ticks.x = element_blank(),
        legend.position = "top") +
  labs(y = "", x = "")


# p1 clustered
all_info |> 
  head(30) |> 
  mutate(Protein = factor(Protein, levels = rev(Protein)),
         Secreted_blood = ifelse(`Secretome location` == "Secreted to blood", 1, 0),
         Pancreatic_cancer_databse = ifelse(Pancreatic_cancer_databse == "Yes", 1, 0),
         Cancer_related = ifelse(Cancer_related == "Yes", 1, 0),
         Enriched_pancreas = ifelse(Enriched_pancreas == "Yes", 1, 0)) |> 
  select(Protein, PC_database = Pancreatic_cancer_databse, Cancer_related,Secreted_blood) |>
  column_to_rownames("Protein") |> 
  pheatmap(clustering_method = "ward.D2", color = c("grey", "red")) |> 
  as.ggplot()


ggsave(savepath("clustered_proteins.png"))
p2 <- 
  all_info |> 
  head(30) |> 
  mutate(Protein = factor(Protein, levels = rev(Protein))) |> 
  select(Protein, `Gene description`, 
         Secretome = `Secretome location`, 
         Specificity = `RNA tissue specificity`) |> 
  pivot_longer(-c(1:2), names_to = "Type", values_to = "Status") |> 
  arrange(Protein) |> 
  ggplot(aes(Type, Protein, fill = Status)) +
  geom_tile(color = "white") +
  scale_fill_manual(values = c(pal_secreted, pal_specificity)) +
  coord_equal() +
  theme_hpa(angled = T) +
  theme(axis.ticks.y = element_blank(),
        axis.line.y = element_blank(),
        axis.text.y = element_blank(),
        axis.ticks.x = element_blank(),
        legend.position = "top") +
  labs(y = "", x = "")

p3 <- 
  all_info |> 
  head(30) |> 
  mutate(Protein = factor(Protein, levels = rev(Protein))) |> 
  select(Protein, `Gene description`, 
         `Random forest` = Importance_rf, 
         Lasso = Importance_lasso) |> 
  pivot_longer(-c(1:2), names_to = "Type", values_to = "Status") |> 
  arrange(Protein) |> 
  ggplot(aes(Type, Protein, fill = Status)) +
  geom_tile(color = "white") +
  scale_fill_viridis_c() +
  coord_equal() +
  theme_hpa(angled = T) +
  theme(axis.ticks.y = element_blank(),
        axis.line.y = element_blank(),
        axis.text.y = element_blank(),
        axis.ticks.x = element_blank(),
        legend.position = "top") +
  labs(y = "", x = "")

p4 <- 
  all_info |> 
  head(30) |> 
  mutate(Protein = factor(Protein, levels = rev(Protein))) |> 
  select(Protein, `Gene description`, 
         `Healthy` = logFC_healthy, 
         Cancers = logFC_cancers) |> 
  pivot_longer(-c(1:2), names_to = "Type", values_to = "Status") |> 
  arrange(Protein) |> 
  ggplot(aes(Type, Protein, fill = Status)) +
  geom_tile(color = "white") +
  scale_fill_gradient(low = "blue", high = "red") +
  #scale_fill_viridis_c() +
  coord_equal() +
  theme_hpa(angled = T) +
  theme(axis.ticks.y = element_blank(),
        axis.line.y = element_blank(),
        axis.text.y = element_blank(),
        axis.ticks.x = element_blank(),
        legend.position = "top") +
  labs(y = "", x = "")


p0 | p1 | p2 | p3  | p4

ggsave(savepath("summary.pdf"), h = 8, w = 8)

p0 | p3 | p4
ggsave(savepath("summary.pdf"), h = 8, w = 8)


# Tissue
consensus_colors <- 
  read_tsv("data/hpa/colors_consensus.tsv")

consensus_colors <-
  consensus_colors %>% 
  mutate(color = ifelse(color == "#DE6C7D", "#B38C6D", color),
         color = ifelse(color == "#A1A8AA", "#DE6C7D", color)) 

pal_tissue <- 
  consensus_colors |> 
  filter(dataset_id == "tissue") |> 
  select(-1) |> 
  bind_rows(tibble(sample = c("amygdala", "appendix", "basal ganglia",
                     "cerebellum","cerebral cortex","colon",
                     "duodenum","hippocampal formation", "hypothalamus",
                     "lymph node","midbrain","rectum"   ,
                     "small intestine","spinal cord","spleen",
                     "thymus","tonsil"), 
                   color = c("#FFDD00","#DE6C7D","#FFDD00",
                     "#FFDD00","#FFDD00", "#1280C4",
                     "#1280C4", "#FFDD00","#FFDD00",
                     "#DE6C7D","#FFDD00", "#1280C4",
                     "#1280C4","#FFDD00", "#DE6C7D",
                      "#DE6C7D", "#DE6C7D"))) |> 
  deframe()

hpa_rna |> 
  distinct(Tissue) |> 
  filter(!Tissue %in% names(pal_tissue))
p2 <- 
  hpa_rna |> 
  filter(`Gene name` %in% proteins) |> 
  group_by(`Gene name`) |> 
  mutate(sum_nTPM = sum(nTPM),
         `Gene name` = factor(`Gene name`, levels = rev(proteins))) |>
  ungroup() |> 
  mutate(nTPM = nTPM/sum_nTPM,
         Tissue = ifelse(nTPM < 0.05, "Other", Tissue),
         Tissue = factor(Tissue, levels = rev(c(names(pal_tissue), "Other")))) |> 
  #filter(nTPM > 0.05) |> 
  ggplot(aes(x = `Gene name`, y = nTPM, fill = Tissue)) +
  geom_col(position = "stack", show.legend = F) +
  scale_fill_manual(values = c(pal_tissue, "Other" = "grey80")) +
  coord_flip() +
  theme_hpa(angled = T) +
  theme(axis.line.y = element_blank(),
        axis.ticks.y = element_blank(),
        axis.text.y = element_blank()) +
  labs(x = "")

```

## Comparison to single classifier

```{r}
# Extract protein ranking
# cancer_recipe |>  
#   prep() |> 
#   juice(all_predictors()) #|>  
# #  names()
# 
# final_glmnet_fit[[1]]

# Extract preprocessed test data (check number of features)

cancer_recipe_test <- 
  recipe(Disease ~ ., data = cancer_train) |> 
  update_role(DAid, new_role = "id") |> 
  step_normalize(all_numeric()) |> 
  step_corr(all_numeric(), threshold = 0.8) |> 
  step_nzv(all_numeric()) |> 
  step_impute_knn(all_numeric()) |> 
  step_downsample(Disease)

cancer_test_preproc <- 
  cancer_recipe  |> 
  prep()  |> 
  bake(new_data = cancer_test) #|> 
  #juice()

# ROC in test set
proteins <- 
  cancer_test_preproc |> 
  select(-DAid, -Disease) |> 
  names()

auc_proteins_test <-
  map_df(proteins, function(protein) {

  roc_res <- pROC::roc(cancer_test_preproc$Disease, cancer_test_preproc |> pull(protein))

  data.frame(protein = protein,
             auc = round(as.numeric(roc_res$auc), 3))
  })

auc_proteins |>
  arrange(-auc) |>
  head()

auc_proteins |>
  ggplot(aes(auc)) +
  #geom_density()
  geom_histogram()


# Plot all ROC together
top_proteins <- 
  important_proteins |> 
  arrange(-abs(estimate)) |> 
  head(5) |> 
  pull(term)

roc_df_proteins <-
  map_df(top_proteins, function(protein) {

  roc_res <- pROC::roc(cancer_test_preproc$Disease, cancer_test_preproc |> pull(protein))
  
  tibble(sensitivity = unlist(roc_res[2]), 
         specificity = unlist(roc_res[3]), 
         Type = protein,
         auc = round(as.numeric(roc_res$auc), 3))
  }) 




auc_final_model <- 
  final_glmnet_fit |> 
  collect_predictions() |> 
  mutate(Disease = factor(Disease, levels = c("Pancreatic cancer", "Healthy"))) |> 
  roc_auc(Disease, `.pred_Pancreatic cancer`) |> 
  pull(.estimate) |> 
  round(3)


final_model_roc <- 
  final_glmnet_fit |> 
  collect_predictions() |> 
  mutate(Disease = factor(Disease, levels = c("Pancreatic cancer", "Healthy"))) |> 
  roc_curve(Disease, `.pred_Pancreatic cancer`) |>
  select(specificity, sensitivity) |> 
  mutate(auc = auc_final_model,
         Type = "Final model (24 proteins)")


library(viridis)
 tibble(sensitivity = unlist(roc_res[2]), 
         specificity = unlist(roc_res[3]), 
         Type = protein,
         auc = round(as.numeric(roc_res$auc), 3))
 
roc_df_proteins |> 
  left_join(enframe(top_proteins), by = c("Type" = "value")) |> 
  mutate(Type = paste0(Type, " (top ", name, ")")) |> 
  select(-name) |>
  bind_rows(final_model_roc) |> 
  mutate(x = 1-specificity,
         Protein = paste(Type, ", AUC: ", auc, sep = ""),
         Protein = fct_reorder(Protein,-auc)) %>% 
    arrange(sensitivity) %>% 
    ggplot(aes(x, sensitivity, group = Protein, color = Protein)) + 
    geom_line(size = 1) + 
    geom_abline(intercept=0, slope=1, linetype="dashed") +
    xlab("1-Specificity") + 
    ylab("Sensitivity") +
    theme_bw() +
    xlab("1-Specificity") + 
  ylab("Sensitivity") +
  scale_color_brewer(palette = "Paired") +
#  scale_color_viridis() +
  scale_x_continuous(breaks = seq(0, 1, 1)) +
  scale_y_continuous(breaks = seq(0, 1, 1)) +
  theme_classic() +
  theme(axis.text = element_text(size = 8))


ggsave(savepath("top_prots_auc.pdf"), h = 4, w = 6)
# Plot ranking against individual AUC
important_proteins |> 
  left_join(auc_proteins, by = c("term" = "protein")) |> 
  ggplot(aes(abs(estimate), auc)) +
  geom_point()

```

## FC across datasets

```{r}

proteins <- 
  all_info |> 
  head(30) |> 
  pull(Protein)

de_healthy <- 
  de |> 
  select(Assay, logFC = logFC_healthy,
         adj.P.Val = adjpval_healthy,
         sig = sig_healthy) |> 
  mutate(Type = "HPA: healthy")

de_cancers <- 
  de |> 
  select(Assay, logFC = logFC_cancers,
         adj.P.Val = adjpval_cancers,
         sig = sig_cancers)|> 
  mutate(Type = "HPA: cancers")

de_all <- 
  de_healthy |> 
  bind_rows(de_cancers) |> 
  bind_rows(de_cptac |> 
  select(Assay, logFC, adj.P.Val, sig) |> 
  mutate(Type = "CPTAC")) |> 
  bind_rows(de_results_ukb |>
  select(Assay, logFC, adj.P.Val, sig) |>
  mutate(Type = "UKB")) 
  # bind_rows(de_results_ukb_1 |> 
  # select(Assay, logFC, adj.P.Val, sig) |> 
  # mutate(Type = "UKB: 1 year before/after")) |> 
  # bind_rows(de_results_ukb_3 |> 
  # select(Assay, logFC, adj.P.Val, sig) |> 
  # mutate(Type = "UKB: 3-5 years")) |> 
  # bind_rows(de_results_ukb_5 |> 
  # select(Assay, logFC, adj.P.Val, sig) |> 
  # mutate(Type = "UKB: 5-7 years"))

fake_scico <- brewer.pal(name = "RdBu", n = 7)

p5 <- 
  de_all |> 
  filter(Assay %in% proteins) |> 
  mutate(Assay = factor(Assay, levels = rev(proteins)),
         Type = factor(Type, levels = c("HPA: healthy",
                                         "HPA: cancers",
                                         "UKB",
                                         "CPTAC"))) |> 
  ggplot(aes(Type, Assay, fill = logFC)) +
  geom_point(aes(size = -log10(adj.P.Val)), color = "black", shape = 21) +
  theme_hpa(angle = T) +
    scale_fill_gradient2(low = fake_scico[7], mid = fake_scico[4], high = fake_scico[1]) +
  scale_size_continuous(range = c(1,3)) +

#  scale_fill_distiller(palette = "RdBu") +
  labs(x = "", y = "") +
  theme(axis.text.y = element_blank(),
        axis.ticks = element_blank(),
        axis.line.y = element_blank(),
        legend.position = "top")

##
de_cptac |> 
  filter(Assay %in% important_proteins_da$term)



library(ggforce)
de_combined |> 
  filter(Assay %in% important_proteins_da$term) |> 
  select(Assay, logFC_cancers, logFC_healthy) |> 
  pivot_longer(cols = c(2:3), names_to = "Comparison", values_to = "logFC") |> 
  mutate(Comparison = recode(Comparison,
                             logFC_cancers = "Cancers",
                             logFC_healthy = "Healthy")) |> 
  ggplot(aes(x = Comparison, y = Assay, fill = logFC)) +
  geom_tile() +
  scale_fill_continuous(type = "gradient") +
#  scale_fill_manual(values = brewer.pal(n = 8, name = "RdBu")) +
  coord_polar(theta = "y")    

#Proteins in all samples

```




# Overlap tissue / blood

```{r}

blood <- 
  de |> 
  filter(sig_healthy != "not significant",
         Assay %in% de_cptac$Assay)

tissue <- 
  de_cptac |> 
  filter(Assay %in% de$Assay,
         sig != "not significant") 


y <- list("Tissue" = unique(tissue |> pull(Assay)),
          "Plasma" = unique(blood |> pull(Assay)))
plot(euler(y, shape = "ellipse"), 
     quantities = TRUE,
     fills = c("#CDCADA", "#E9C2A6")) |> as.ggplot()

ggsave(savepath("overlap_deregulated.pdf"), h = 3, w = 3)

blood |> 
  filter(Assay %in% tissue$Assay)

blood |> 
  filter(!Assay %in% tissue$Assay)

tissue |> 
  filter(!Assay %in% blood$Assay)

# Overlapping proteins

pcd_all |> head()
blood |> 
  filter(Assay %in% tissue$Assay, 
         !Assay %in% pcd_all$Protein)


protein <- "NDUFS6"
plot_protein_datsets(protein = protein, y_axis = T, x_axis = F, title = T)
ggsave(savepath(paste(protein, "_boxplot_all_datasets.png")), h = 4, w = 8)


# Study only tissue
tissue_functions <- 
  tissue |> 
  filter(!Assay %in% blood$Assay) |> 
  left_join(hpa_info_short, by = c("Assay" = "Gene")) |>
  select(Assay, `Protein class`) |> 
  separate_rows(`Protein class`, sep = ", ") |> 
  group_by(`Protein class`) |> 
  summarize(n = n()) |> 
  arrange(-n) |> 
  ggplot(aes(fct_reorder(`Protein class`, n), n)) +
  geom_col() +
  geom_text(aes(label = n, y = n + 1)) +
  theme_hpa(angled = T) +
  ggtitle("Deregulated in tissue") +
  xlab("") +
  coord_flip()


# Study only plasma
plasma_functions <- 
  blood |> 
  filter(!Assay %in% tissue$Assay) |> 
  left_join(hpa_info_short, by = c("Assay" = "Gene")) |>
  select(Assay, `Protein class`) |> 
  separate_rows(`Protein class`, sep = ", ") |> 
  group_by(`Protein class`) |> 
  summarize(n = n()) |> 
  arrange(-n) |> 
  ggplot(aes(fct_reorder(`Protein class`, n), n)) +
  geom_col() +
  geom_text(aes(label = n, y = n + 1)) +
  theme_hpa(angled = T) +
  ggtitle("Deregulated in plasma") +
  xlab("") +
  coord_flip()


# Study overlap
both_functions <- 
  blood |> 
  filter(Assay %in% tissue$Assay) |> 
  left_join(hpa_info_short, by = c("Assay" = "Gene")) |>
  select(Assay, `Protein class`) |> 
  separate_rows(`Protein class`, sep = ", ") |> 
  group_by(`Protein class`) |> 
  summarize(n = n()) |> 
  arrange(-n) |> 
  ggplot(aes(fct_reorder(`Protein class`, n), n)) +
  geom_col() +
  geom_text(aes(label = n, y = n + 1)) +
  theme_hpa(angled = T) +
  ggtitle("Deregulated in both") +
  xlab("") +
  coord_flip()


tissue_functions | plasma_functions | both_functions
tissue_functions / plasma_functions / both_functions


# Proportion

tissue_functions <- 
  tissue |> 
  filter(!Assay %in% blood$Assay) |> 
  left_join(hpa_info_short, by = c("Assay" = "Gene")) |>
  select(Assay, `Protein class`) |> 
  separate_rows(`Protein class`, sep = ", ") |> 
  group_by(`Protein class`)  |> 
  summarize(perc = n()/52 * 100, n = n()) |> 
  arrange(-n) |> 
  ggplot(aes(fct_reorder(`Protein class`, n), perc)) +
  geom_col() +
  scale_y_continuous(limits = c(0,100)) +
  geom_text(aes(label = n, y = perc + 5)) +
  theme_hpa(angled = T) +
  ggtitle("Deregulated in tissue (n = 52)") +
  xlab("") +
  ylab("% of proteins") +
  coord_flip()


# Study only plasma
plasma_functions <- 
  blood |> 
  filter(!Assay %in% tissue$Assay) |> 
  left_join(hpa_info_short, by = c("Assay" = "Gene")) |>
  select(Assay, `Protein class`) |> 
  separate_rows(`Protein class`, sep = ", ") |> 
  group_by(`Protein class`)  |> 
  summarize(perc = n()/114 * 100, n = n()) |> 
  arrange(-n) |> 
  ggplot(aes(fct_reorder(`Protein class`, n), perc)) +
  geom_col() +
  scale_y_continuous(limits = c(0,100)) +
  geom_text(aes(label = n, y = perc + 5)) +
  theme_hpa(angled = T) +
  ggtitle("Deregulated in plasma (n = 114)") +
  xlab("") +
  ylab("% of proteins") +
  coord_flip()


# Study overlap
both_functions <- 
  blood |> 
  filter(Assay %in% tissue$Assay) |> 
  left_join(hpa_info_short, by = c("Assay" = "Gene")) |>
  select(Assay, `Protein class`) |> 
  separate_rows(`Protein class`, sep = ", ") |> 
  group_by(`Protein class`) |> 
  summarize(perc = n()/22 * 100, n = n()) |> 
  arrange(-n) |> 
  ggplot(aes(fct_reorder(`Protein class`, n), perc)) +
  geom_col() +
  scale_y_continuous(limits = c(0,100)) +
  geom_text(aes(label = n, y = perc + 5)) +
  theme_hpa(angled = T) +
  ggtitle("Deregulated in both (n = 22)") +
  xlab("") +
  ylab("% of proteins") +
  coord_flip()

tissue_functions / plasma_functions / both_functions

ggsave(savepath("overview_percentage.png"), h = 12, w = 8)
```

## Enrichment

```{r}

library(clusterProfiler)
original_gene_list <- 
  blood |> 
  filter(!Assay %in% tissue$Assay) |> 
  left_join(cancer_data_long |> 
              distinct(Assay, UniProt), by = "Assay") |> 
  select(UniProt, logFC_healthy) |> 
  deframe()

gene_list <-  sort(original_gene_list, decreasing = TRUE)

kegg_organism <-  "hsa"

plasma_kegg_res <- gseKEGG(geneList     = gene_list,
               organism     = kegg_organism,
               #nPerm        = 10000,
               #minGSSize    = 3,
               #maxGSSize    = 800,
               pvalueCutoff = 0.05,
               pAdjustMethod = "BH",
               keyType       = "uniprot")

plasma_kegg_res

```





# Protein report


```{r}

protein <- "MDK"
protein_summary(protein)


protein_plots <- map(top_30, ~protein_summary(.))

pdf(savepath("top2.pdf"), w = 10, h = 15)
protein_plots
dev.off()


protein <-  "BAIAP2"
pdf(savepath(paste0(protein, "_summary.png")), w = 10, h = 15)
protein_summary(protein)
dev.off()
```


# Summary top 200

```{r}

rank_uniprot <- 
  all_rank |> 
  left_join(cancer_data_long |> 
              distinct(Assay, UniProt), by = "Assay")

top_200 <- 
  rank_uniprot |> 
  head(100) |> 
  pull(UniProt)

go_enrich <- enrichGO(gene = top_200,
                      universe = rank_uniprot$UniProt,
                      OrgDb = "org.Hs.eg.db", 
                      keyType = 'UNIPROT',
                      readable = T,
                      ont = "BP",
                      pvalueCutoff = 0.05, 
                      qvalueCutoff = 0.10)

cnetplot(go_enrich) 

go_enrich |> 
  as_tibble()
```

# Summary top 30

```{r}
GO_terms |> 
  left_join(hpa_protein_class |> 
              distinct(display_name, ensg_id) |> 
              rename(protein = display_name, `Gene stable ID` = ensg_id), by = "Gene stable ID") |> 
  filter(protein == "ENPP5",
         `GO domain` == "biological_process")

top_30

hpa_protein_class |> 
  filter(display_name %in% top_30)

mapping <- 
  hpa_info_short |> 
  distinct(protein = Gene,
           gene = Ensembl)

top_30_ensembl <- 
  mapping |> 
  filter(protein %in% top_30) |> 
  pull(gene)
 
dat_class <- 
  hpa_info_short |> 
  filter(Gene %in% top_30) |> 
  select(Gene, `Protein class`) |> 
  separate_rows(`Protein class`, sep = ", ") |> 
  mutate(value = 1) |> 
  pivot_wider(names_from = `Protein class`, values_from = "value", values_fill = 0) |> 
  pivot_longer(cols = c(2:16), names_to = "Functions", values_to = "Value") |> 
  mutate(Value = factor(Value)) |> 
  mutate(Gene = factor(Gene, levels = rev(top_30))) 

class_order <- 
  dat_class |> 
  filter(Value == 1) |> 
  group_by(Functions) |> 
  summarize(n = n()) |> 
  arrange(-n)

plot_class <- 
  dat_class |> 
  mutate(Functions = factor(Functions, levels = class_order$Functions)) |> 
  ggplot(aes(Functions, Gene, fill = Value)) +
  geom_tile(show.legend = F) +
  coord_fixed() +
  scale_fill_manual(values = c("grey90", "darkred")) +
  theme_hpa(angled = T) +
  xlab("") +
  ggtitle("Protein class") 


# dat_tissue <- 
#   hpa_enhanced_tissues |> 
#   filter(ensg_id %in% top_30_ensembl) |> 
#   select(Gene = ensg_id, enhanced_tissues) |> 
#   separate_rows("enhanced_tissues", sep = ",") |> 
#   mutate(value = 1) |> 
#   rename(Tissues = enhanced_tissues) |> 
#   pivot_wider(names_from = "Tissues", values_from = "value", values_fill = 0) |> 
#   pivot_longer(cols = c(2:22), names_to = "Tissues", values_to = "Value") |> 
#   mutate(Value = factor(Value)) |> 
#   mutate(Gene = factor(Gene, levels = rev(top_30_ensembl))) 
# 
# tissue_order <- 
#   dat_tissue |> 
#   filter(Value == 1) |> 
#   group_by(Tissues) |> 
#   summarize(n = n()) |> 
#   arrange(-n)
# 
# dat_tissue |> 
#   mutate(Tissues = factor(Tissues, levels = tissue_order$Tissues)) |> 
#   ggplot(aes(Tissues, Gene, fill = Value)) +
#   geom_tile() +
#   coord_fixed() +
#   scale_fill_manual(values = c("grey90", "darkred")) +
#   theme_hpa(angled = T) +
#   ggtitle("Tissue enhanced")


go_dat <- 
  GO_terms |> 
  rename(Gene = `Gene stable ID`) |> 
  filter(Gene %in% top_30_ensembl,
         `GO domain` == "biological_process") |> 
  distinct()  

go_order <- 
  go_dat |>
  group_by(`GO term name`) |> 
  summarise(n = n()) |>
  arrange(-n) |> 
  head(15)

plot_go <- 
  tibble(Gene = top_30_ensembl) |> 
  left_join(go_dat|> 
                filter(`GO term name` %in% go_order$`GO term name`) |> 
                mutate(value = 1) |> 
                select(Gene, GO = `GO term name`, value), by = "Gene") |> 
  mutate(value = ifelse(is.na(value), 0, value)) |> 
  pivot_wider(names_from = "GO", values_from = "value", values_fill = 0) |> 
  select(-`NA`) |> 
  pivot_longer(cols = c(2:16), names_to = "GO", values_to = "Value") |> 
  mutate(Value = factor(Value)) |> 
  left_join(mapping, by = c("Gene" = "gene")) |> 
  select(-Gene) |> 
  rename(Gene = protein) |> 
  mutate(Gene = factor(Gene, levels = rev(top_30)),
         GO = factor(GO, levels = go_order$`GO term name`))  |> 
  ggplot(aes(GO, Gene, fill = Value)) +
  geom_tile(show.legend = F) +
  coord_fixed() +
  scale_fill_manual(values = c("grey90", "darkred")) +
  theme_hpa(angled = T, axis_y = F) +
  xlab("") +
  ggtitle("GO terms")


reactome_dat <- 
  reactome |> 
  rename(Gene = ensg_id) |> 
  filter(Gene %in% top_30_ensembl) |> 
  distinct()  

reactome_order <- 
  reactome_dat |>
  group_by(description) |> 
  summarise(n = n()) |>
  arrange(-n) |> 
  head(15)

plot_reactome <- 
  tibble(Gene = top_30_ensembl) |> 
  left_join(reactome_dat |> 
              filter(description %in% reactome_order$description) |> 
            mutate(value = 1) |> 
              select(Gene, Reactome = description, value) |> 
              distinct(), by = "Gene") |> 
  mutate(value = ifelse(is.na(value), 0, value)) |> 
  pivot_wider(names_from = "Reactome", values_from = "value", values_fill = 0) |> 
  select(-`NA`) |> 
  pivot_longer(cols = c(2:16), names_to = "Reactome", values_to = "Value") |> 
  mutate(Value = factor(Value)) |> 
  left_join(mapping, by = c("Gene" = "gene")) |> 
  select(-Gene) |> 
  rename(Gene = protein) |>
  mutate(Gene = factor(Gene, levels = rev(top_30)),
         Reactome = factor(Reactome, levels = reactome_order$description)) |>
  ggplot(aes(Reactome, Gene, fill = Value)) +
  geom_tile() +
  coord_fixed() +
  scale_fill_manual(values = c("grey90", "darkred")) +
  theme_hpa(angled = T, axis_y = F) +
  xlab("") +
  ggtitle("Reactome terms")


plot_class | plot_go | plot_reactome
ggsave(savepath("sumamry_classes.png"), h = 12, w = 14)

```

## Clustering

```{r}

all_dat <- 
  dat_class |> 
  rename(Term = Functions) |> 
  mutate(Type = "Class") |> 
  bind_rows(tibble(Gene = top_30_ensembl) |> 
              left_join(go_dat|> 
                          filter(`GO term name` %in% go_order$`GO term name`) |> 
                          mutate(value = 1) |> 
                          select(Gene, GO = `GO term name`, value), by = "Gene") |> 
              mutate(value = ifelse(is.na(value), 0, value)) |> 
              pivot_wider(names_from = "GO", values_from = "value", values_fill = 0) |> 
              select(-`NA`) |> 
              pivot_longer(cols = c(2:16), names_to = "GO", values_to = "Value") |> 
              mutate(Value = factor(Value)) |> 
              left_join(mapping, by = c("Gene" = "gene")) |> 
              select(-Gene) |> 
              rename(Gene = protein) |> 
              mutate(Gene = factor(Gene, levels = rev(top_30)),
                     GO = factor(GO, levels = go_order$`GO term name`)) |> 
              rename(Term = GO) |> 
              mutate(Type = "GO")) |> 
  bind_rows( tibble(Gene = top_30_ensembl) |> 
               left_join(reactome_dat |> 
                           filter(description %in% reactome_order$description) |> 
                           mutate(value = 1) |> 
                           select(Gene, Reactome = description, value) |> 
                           distinct(), by = "Gene") |> 
               mutate(value = ifelse(is.na(value), 0, value)) |> 
               pivot_wider(names_from = "Reactome", values_from = "value", values_fill = 0) |> 
               select(-`NA`) |> 
               pivot_longer(cols = c(2:16), names_to = "Reactome", values_to = "Value") |> 
               mutate(Value = factor(Value)) |> 
               left_join(mapping, by = c("Gene" = "gene")) |> 
               select(-Gene) |> 
               rename(Gene = protein) |>
               mutate(Gene = factor(Gene, levels = rev(top_30)),
                      Reactome = factor(Reactome, levels = reactome_order$description)) |> 
               rename(Term = Reactome) |> 
               mutate(Type = "Reactome"))


ann <- 
  all_dat |> 
  distinct(Term, Type) |>
  column_to_rownames("Term")

ann_row <- 
  tibble(Gene = top_30) |> 
  mutate(In_PCD = ifelse(Gene %in% pcd_all$Protein, "Yes", "No")) |> 
  column_to_rownames("Gene")

all_dat |> 
  select(-Type) |> 
  mutate(Value = as.numeric(Value)) |> 
  pivot_wider(names_from = "Term", values_from = "Value") |> 
  column_to_rownames("Gene") |> 
  pheatmap(clustering_method = "ward.D2",
           color = c("grey80", "darkred"),
           border_color = "white",
           cutree_rows = 6,
           annotation_col = ann,
           annotation_row = ann_row,
           annotation_colors = list("In_PCD" = c("No" = "darkblue", 
                                                 "Yes" = "red"))) |> 
  as.ggplot()

ggsave(savepath("clustering_genes.png"), h = 12, w = 12)

heat_dat <- 
    all_dat |> 
  select(-Type) |> 
  mutate(Value = as.numeric(Value)) |> 
  pivot_wider(names_from = "Term", values_from = "Value") |> 
  column_to_rownames("Gene") 

p <- 
  heat_dat |> 
  pheatmap(clustering_method = "ward.D2",
           color = c("grey80", "darkred"),
           border_color = "white",
           cutree_rows = 6,
           annotation_col = ann,
           annotation_row = ann_row,
           annotation_colors = list("In_PCD" = c("No" = "darkblue", 
                                                 "Yes" = "red"))) 
  
clustered_top30 <- rownames(heat_dat[p$tree_row[["order"]],])

p |> 
  as.ggplot()

ggsave(savepath("clustering_terms.pdf"), h = 8, w = 8)

```

#### Add rest of plot

```{r}
all_info <- 
  hpa_info_short |> 
  select(Protein = Gene, 
         `Gene description`, 
         `Protein class`, 
         `RNA tissue specificity`,
         `RNA tissue specific nTPM`,
         `Secretome location`) |> 
  filter(Protein %in% all_rank$Assay) |> 
  mutate(Protein = factor(Protein, levels = all_rank$Assay)) |> 
  arrange(Protein) |> 
  left_join(all_rank |> 
              select(-rank_de_cancers, -rank_de_healthy, -rank_rf, -rank_lasso, -sum_rank, -Significance, -Type) |> 
              rename(Protein = Assay), by = "Protein") |> 
  mutate(Cancer_related = ifelse(grepl("Cancer-related", `Protein class`), "Yes", "No"),
         Enriched_pancreas = ifelse(`RNA tissue specificity` %in% c("Tissue enhanced", "Tissue enriched") & grepl("pancreas", `RNA tissue specific nTPM`), "Yes", "No")) |> 
  select(-`Protein class`) |> 
  mutate(Pancreatic_cancer_databse = ifelse(Protein %in% pcd_proteins, "Yes", "No"))

p0 <- 
  all_info |> 
  head(30) |> 
  mutate(Protein = factor(Protein, levels = clustered_top30)) |> 
  arrange(Protein) |> 
  mutate(`Gene description` = factor(`Gene description`, levels = rev(`Gene description`))) |> 
  select(Protein, `Gene description`) |> 
  arrange(Protein) |> 
  ggplot(aes(x = "", y = `Gene description`)) +
  geom_text(aes(label = Protein), hjust = 0.5, fontface = "bold", size = 3) +
  theme_void() +
  theme(axis.text.y = element_text(hjust = 1, size = 8))


p1 <- 
  all_info |> 
  head(30) |> 
  mutate(Protein = factor(Protein, levels = rev(clustered_top30)),
         Secreted_blood = ifelse(`Secretome location` == "Secreted to blood", "Yes", "No")) |> 
  select(Protein, `Gene description`, PC_database = Pancreatic_cancer_databse, Cancer_related,Secreted_blood,
         Enriched_pancreas) |>
  pivot_longer(-c(1:2), names_to = "Type", values_to = "Status") |> 
  mutate(Type = factor(Type, levels = c("PC_database",
                                        "Cancer_related",
                                        "Secreted_blood",
                                        "Enriched_pancreas"))) |> 
  arrange(Protein) |> 
  ggplot(aes(Type, Protein, fill = Status)) +
  geom_tile(color = "white") +
  scale_fill_manual(values = c("grey", "red")) +
  coord_equal() +
  theme_hpa(angled = T) +
  theme(axis.ticks.y = element_blank(),
        axis.line.y = element_blank(),
        axis.text.y = element_blank(),
        axis.ticks.x = element_blank(),
        legend.position = "top") +
  labs(y = "", x = "")

p3 <- 
  all_info |> 
  head(30) |> 
  mutate(Protein = factor(Protein, levels = rev(clustered_top30))) |> 
  select(Protein, `Gene description`, 
         `Random forest` = Importance_rf, 
         Lasso = Importance_lasso) |> 
  pivot_longer(-c(1:2), names_to = "Type", values_to = "Status") |> 
  arrange(Protein) |> 
  ggplot(aes(Type, Protein, fill = Status)) +
  geom_tile(color = "white") +
  scale_fill_viridis_c() +
  coord_equal() +
  theme_hpa(angled = T) +
  theme(axis.ticks.y = element_blank(),
        axis.line.y = element_blank(),
        axis.text.y = element_blank(),
        axis.ticks.x = element_blank(),
        legend.position = "top") +
  labs(y = "", x = "")

p4 <- 
  all_info |> 
  head(30) |> 
  mutate(Protein = factor(Protein, levels = rev(clustered_top30))) |> 
  select(Protein, `Gene description`, 
         `Healthy` = logFC_healthy, 
         Cancers = logFC_cancers) |> 
  pivot_longer(-c(1:2), names_to = "Type", values_to = "Status") |> 
  arrange(Protein) |> 
  ggplot(aes(Type, Protein, fill = Status)) +
  geom_tile(color = "white") +
  scale_fill_gradient(low = "blue", high = "red") +
  #scale_fill_viridis_c() +
  coord_equal() +
  theme_hpa(angled = T) +
  theme(axis.ticks.y = element_blank(),
        axis.line.y = element_blank(),
        axis.text.y = element_blank(),
        axis.ticks.x = element_blank(),
        legend.position = "top") +
  labs(y = "", x = "")

p2 <- 
  hpa_rna |> 
  filter(`Gene name` %in% top_30) |> 
  group_by(`Gene name`) |> 
  mutate(sum_nTPM = sum(nTPM),
         `Gene name` = factor(`Gene name`, levels = rev(clustered_top30))) |>
  ungroup() |> 
  mutate(nTPM = nTPM/sum_nTPM,
         Tissue = ifelse(nTPM < 0.05, "Other", Tissue),
         Tissue = factor(Tissue, levels = rev(c(names(pal_tissue), "Other")))) |> 
  #filter(nTPM > 0.05) |> 
  ggplot(aes(x = `Gene name`, y = nTPM, fill = Tissue)) +
  geom_col(position = "stack", show.legend = F) +
  scale_fill_manual(values = c(pal_tissue, "Other" = "grey80")) +
  coord_flip() +
  theme_hpa(angled = T) +
  theme(axis.line.y = element_blank(),
        axis.ticks.y = element_blank(),
        axis.text.y = element_blank()) +
  labs(x = "")

p5 <- p |> as.ggplot()

p0 | p1 | p2 | p3  | p4 

ggsave(savepath("summary_clustered.pdf"), h = 8, w = 8)
```



## Graph

```{r}

all_dat

library(tidygraph)
library(ggraph)

# Assuming your tibble is called `gene_info`
# Make sure to replace this with your actual tibble name

secreted_prots <- 
  hpa_info_short |> 
  filter(`Secretome location`  == "Secreted to blood")

# REACTOME
gene_info <- 
  all_dat |> 
  mutate(Secreted = ifelse(Gene %in% secreted_prots$Gene, "Yes", "No"),
         In_PCD = ifelse(Gene %in% pcd_all$Protein, "Yes", "No")) |> 
  filter(Type == "Reactome")

dat <- 
  gene_info %>%
  filter(Value == 1) %>%
  select(from = Gene, to = Term, Secreted, In_PCD, Type) |> 
  as_tbl_graph(directed = FALSE) |> 
  activate(nodes) %>%
  mutate(community = as.factor(group_louvain()),
         Node_type = ifelse(name %in% top_30, 10000, 0)) 

set.seed(213)
dat %>% 
  ggraph(layout = "kk") +
  geom_edge_link(alpha = 0.8, color = "grey90") +
  geom_node_point(aes(colour = community, size = Node_type), size = 10) + 
  geom_node_text(aes(label = name, color = community), color = "white", size = 2.5, fontface = "bold") +
#  geom_node_text(data = . |> filter(Node_type == 0), aes(label = name, color = community)) +
 # geom_node_text(data = . |> filter(Node_type == 10000), aes(label = name), color = "black") +
  scale_color_manual(values = c("#6d597a", "#b56576", "#e56b6f", "#eaac8b")) +
  scale_fill_manual(values = c("#6d597a", "#b56576", "#e56b6f", "#eaac8b")) + 
  theme_graph() +
  ggtitle("Reactome")

ggsave(savepath("reactome_network.pdf"), h = 10, w = 10)
 


# GENE ONTOLOGY

secreted_prots <- 
  hpa_info_short |> 
  filter(`Secretome location`  == "Secreted to blood")

# REACTOME
gene_info <- 
  all_dat |> 
  mutate(Secreted = ifelse(Gene %in% secreted_prots$Gene, "Yes", "No"),
         In_PCD = ifelse(Gene %in% pcd_all$Protein, "Yes", "No")) |> 
  filter(Type == "Class")

dat <- 
  gene_info %>%
  filter(Value == 1) %>%
  select(from = Gene, to = Term, Secreted, In_PCD, Type) |> 
  as_tbl_graph(directed = FALSE) |> 
  activate(nodes) %>%
  mutate(community = as.factor(group_louvain()),
         Node_type = ifelse(name %in% top_30, 10000, 0)) 

set.seed(213)

layout <- layout_with_drl(graph)

dat %>% 
  ggraph(layout = "kk") +
  geom_edge_link(alpha = 0.8, color = "grey90") +
  geom_node_point(aes(colour = community, size = Node_type), size = 10) + 
  geom_node_text(aes(label = name, color = community), color = "black", size = 2.5, fontface = "bold") +
#  geom_node_text(data = . |> filter(Node_type == 0), aes(label = name, color = community)) +
 # geom_node_text(data = . |> filter(Node_type == 10000), aes(label = name), color = "black") +
#  scale_color_manual(values = c("#6d597a", "#b56576", "#e56b6f", "#eaac8b")) +
 # scale_fill_manual(values = c("#6d597a", "#b56576", "#e56b6f", "#eaac8b")) + 
  theme_graph() 

ggsave(savepath("protein_class_network.pdf"), h = 10, w = 10)


# PROTEIN CLASS




# Open cell







# Correlation network
library(corrr)

prots_cor <- 
  cancer_data_long |> 
  filter(Assay %in% top_30,
         Disease %in% c("PAN", "CAD healthy")) |> 
  select(DAid, Assay, NPX) |> 
  pivot_wider(names_from = "Assay", values_from = "NPX") |> 
  column_to_rownames("DAid") |> 
  correlate() |> 
  shave(upper = T) |> 
  stretch(na.rm = T) |>   
  mutate(r = abs(r)) |> 
  filter(r >= 0.7) 

# Positive or negative correlation
# Protein type (scatter pie)
# Secreted (border)
# Add singlets
library(scatterpie)

node_info_long <- 
  hpa_info_short |>
  select(Gene, class = `Protein class`) |> 
   separate_rows(class, sep = ", ") |> 
   filter(Gene %in% top_30) |> 
   mutate(value = 1) |> 
   group_by(Gene) |> 
   mutate(value = value / n()) 

node_info <- 
  node_info_long |> 
   pivot_wider(names_from = class, values_from = value, values_fill = 0)
 
set.seed(100)
cor_graph <- as_tbl_graph(prots_cor, directed = FALSE)

cor_graph_ext <- 
  cor_graph |> 
  activate(nodes) |> 
  left_join(node_info, by = c("name" = "Gene")) 

#Precompute layout
#precompute the layout
library(ggforce)
library(graphlayouts)

xy <- layout_with_stress(cor_graph_ext)
cor_graph_ext$x <- xy[, 1]
cor_graph_ext$y <- xy[, 2]


cor_layout <- 
  cor_graph_ext |> 
  create_layout("kk") 

scatter <- 
  cor_layout |> 
  select(x, y, name) |> 
  left_join(node_info, by = c("name" = "Gene"))

columns <- 
  scatter |> 
  as_tibble() |> 
  select(4:18) |> 
  colnames()

library("ggsci")

cor_layout |> 
  #ggraph("manual", x = cor_graph_ext$x, y = cor_graph_ext$y) + 
  ggraph() +
  geom_edge_link(aes(width = r), alpha = 0.2) + 
  scale_edge_width(range = c(0.2, 1.8)) +
  geom_node_point(size = 5) +
  geom_scatterpie(
    cols = columns,
    data = scatter,
    #aes(x, y, fill = class),
    colour = NA,
    pie_scale = 2
  ) +
  scale_fill_manual(values = pal) +
  coord_fixed() +
  #scale_fill_brewer(palette = "tableau_cyclic") +
  #scale_fill_npg() +
  geom_node_text(aes(label = name), size = 3, repel = F, color = "black", fontface = "bold") +
  theme_graph()

ggsave(savepath("net_cor.pdf"))
makeTransparent<-function(someColor, alpha=100)
{
  newColor<-someColor + alpha #I wish
  return(newColor)
}

pal <- 
  c("#296478",
    "#3481B9",
    "#AED9E0",
    "#70A9A1",
    "#9EC1A3",
    "#CFE0C3",
    "#EF3054",
    "#8B0000",
    "#E74929",
    "#E2B4BD",
    "#D3917E",
    "#F19143",
    "#DDB967",
    "#D0E37F",
    "#E3E7D3")

pal_transparent <- map(pal, function(color) { adjustcolor(color, alpha = 0.8) }) |> unlist()

# Step 1: Create an edge table from your tibble
edges <- 
  gene_info %>%
  filter(Value == 1) %>%
  select(Protein = Gene, Term, Type) 

# Step 2: Create a node table from your tibble
nodes <- 
  gene_info %>%
  rename(Protein = Gene) |> 
  distinct(Protein, .keep_all = TRUE) %>%
  select(Protein, Secreted, In_PCD) |> 
  select(Protein)

# Step 3: Create a tidygraph object
graph <- tbl_graph(nodes = nodes, edges = edges)

# Step 4: Prepare node attributes for visualization
node_attributes <- nodes %>%
  mutate(
    color = ifelse(secreted == 1, "Secreted", "Not Secreted"),
    border = ifelse(in_database == 1, "In Database", "Not in Database")
  )

# Step 5: Create the plot
ggraph(graph, layout = 'fr') +
  geom_edge_link() +
  geom_node_point(aes(color = color, shape = border)) +
  scale_color_manual(values = c("Secreted" = "red", "Not Secreted" = "blue")) +
  scale_shape_manual(values = c("In Database" = 15, "Not in Database" = 16)) +
  geom_node_text(aes(label = name), repel = TRUE) +
  labs(title = "Gene-Term Network",
       subtitle = "Nodes: Genes and Terms",
       caption = "Node color: Secreted status; Node border: Database status") +
  theme_void()



gene_info <- tibble(
  Protein = c("GeneA", "GeneB", "GeneC", "GeneD"),
  Term = c("Term1", "Term2", "Term3", "Term1"),
  Type = c("Class", "GO", "Reactome", "Class"),
  value = c(1, 1, 1, 1),
  secreted = c(1, 0, 1, 0),
  in_database = c(1, 0, 1, 1)
)

# Create an edge table
edges <- gene_info %>%
  filter(value == 1) %>%
  select(Protein, Term)

# Create a node table
nodes <- gene_info %>%
  distinct(Protein, .keep_all = TRUE) %>%
  select(Protein, secreted, in_database)

# Create a tidygraph object
graph <- tbl_graph(nodes = nodes, edges = edges, directed = FALSE)

```

### Open cell

```{r}
# Ribosomal network
opencell <- 
  read_csv("data/hpa/opencell-protein-interactions.csv")


mapping_30 <- 
  hpa_info_short |> 
  filter(Gene %in% top_30) |> 
  select(1,2)

opencell %>% 
  select(3, 4) %>% 
  filter(interactor_ensg_id %in% mapping_30$Ensembl)
  left_join(mapping_30, by = c("target_ensg_id" = "Ensembl")) |> 
  select(-target_ensg_id) |> 
  rename(target_ensg_id = Gene) |> 
   left_join(mapping_30, by = c("interactor_ensg_id" = "Ensembl")) |> 
  select(-interactor_ensg_id) |> 
  rename(interactor_ensg_id = Gene) |> 
  filter(!is.na(target_ensg_id),
         !is.na(interactor_ensg_id))

  left_join(clustering_data %>% 
              filter(dataset_id == "singlecell") %>% 
              select(gene, cluster),
            by = c("target_ensg_id" = "gene")) %>% 
  left_join(clustering_data %>% 
              filter(dataset_id == "singlecell") %>% 
              select(gene, cluster),
            by = c("interactor_ensg_id" = "gene"),
            suffix = c("_target", "_interactor")) %>% 
  set_colnames(c("gene1", "gene2", "cluster1", "cluster2")) %>%
  
  bind_rows(tibble(gene1 = .$gene2, gene2 = .$gene1,
                   cluster1 = .$cluster2, cluster2 = .$cluster1)) %>% 
  filter(cluster1 < cluster2 |
           (cluster1 == cluster2 & 
              gene1 < gene2)) %>% 
  distinct() %>% 
  group_by(cluster1) %>% 
  mutate(tot1 = n_distinct(gene1)) %>% 
  group_by(cluster2) %>% 
  mutate(tot2 = n_distinct(gene2)) %>% 
  group_by(cluster1, cluster2, tot1, tot2) %>% 
  count() %>% 
  ungroup() %>% 
  group_by_all() %>% 
  mutate(frac = n / min(c(tot1, tot2))) %>% 
  ungroup() %>% 
  arrange(-frac) %>% 
  # filter(cluster1 == cluster2) %>% 
  print(n = 100)
# as_tbl_graph()




plot_genes <- 
  clustering_data %>% 
  filter(dataset_id == "singlecell") %>% 
  filter(cluster %in% c(39)) %>% 
  arrange(cluster) %>% 
  left_join(gene_info %>% 
              select(gene = ensg_id,
                     gene_name))


table(plot_genes$gene %in% c(opencell$target_ensg_id, 
                             opencell$interactor_ensg_id))

ribo_prots <- read_tsv("data/meta/ribosomal_proteins.txt")

opencell_39 <- 
  opencell %>% 
  filter(target_ensg_id %in% plot_genes$gene |
           interactor_ensg_id %in% plot_genes$gene) %>% 
  mutate(type = case_when(target_ensg_id %in% plot_genes$gene &
                            interactor_ensg_id %in% plot_genes$gene ~ "both",
                          target_ensg_id %in% plot_genes$gene ~ "target",
                          interactor_ensg_id %in% plot_genes$gene ~ "interactor"),
         type_ribo = case_when(target_ensg_id %in% ribo_prots$ensg_id &
                                 interactor_ensg_id %in% ribo_prots$ensg_id ~ "both",
                               target_ensg_id %in% ribo_prots$ensg_id ~ "target",
                               interactor_ensg_id %in% ribo_prots$ensg_id ~ "interactor")) 

opencell_39 %>% 
  select(3, 4) %>% 
  gather(a, b) %>% 
  select(2) %>% 
  filter(b %in% plot_genes$gene) %>% 
  distinct()



opencell_39 %>% 
  select(1, 2, type) %T>% 
  {
    group_by(., type) %>% 
      count() %>% 
      print()
  } %>% 
  as_tbl_graph() %>% 
  ggraph() +
  geom_edge_link() +
  geom_node_point(aes(color = name %in% plot_genes$gene_name)) +
  coord_fixed() +
  theme_void()

opencell_39 %>% 
  select(3, 4, type) %T>% 
  {
    group_by(., type) %>% 
      count() %>% 
      print()
  } %>% 
  as_tbl_graph() %>% 
  ggraph() +
  geom_edge_link() +
  geom_node_point(aes(color = name %in% ribo_prots$ensg_id)) +
  coord_fixed() +
  theme_void()

opencell_39 %>% 
  filter(type == "both") %>% 
  select(3, 4) %>% 
  gather(a, b) %>% 
  pull(b) %>% 
  n_distinct()

# Figure
plot_data <-
  opencell_39 %>% 
  filter(type == "both") %>%
  select(3, 4, type) %>% 
  distinct() %>% 
  as_tbl_graph() %>% 
  mutate(component = group_components(), 
         degree = centrality_degree()) %>% 
  left_join(ribosomal_data_extended %>% 
              select(name = ensembl_gene_id, gene_name, type = figure_key))

library(ggsci)

levels <- 
  c("Ribosomal proteins", "Protein biosynthesis", "Chaperone", "Other")


#graphopt
#fr

#p2 <- 
set.seed(213)
  plot_data %>% 
  mutate(type = factor(type, levels = levels)) %>% 
  group_by(component) %>% 
  ggraph(layout = "kk") + #kk
  geom_edge_link(color = "gray60", width = 0.25,
                 alpha = 0.15) +
  geom_node_point(aes(color = type),
                  size = 2) +
  # geom_node_text(aes(label = gene_name),
  #                size = 2,
  #                vjust = -0.5) +
  ylim(-10, 10) +
  xlim(-10, 10) +
  coord_fixed() +
  scale_color_manual(values = c("#E53935", "#5ab4ac","#DDC024",  "grey")) +
 # scale_color_simpsons() +
  #scale_color_futurama()+
  theme_void()
#   
# "#E53935" 
# "#808080"
# "#B0BEC5"
# "#A1887F"
# "#AED6F1"
# "#98D98E"
# "#FEDD72"
# "#DB3655"
# "#F5A221"
# "#38BCDE"
# "#583086"
# "#87CFD2"
# "#DDC024"

  
ggsave(savepath("Cluster_39_new_network_notext.pdf"),
       width = 8, height = 8)

```

# Radar plots

```{r}

cptac_data_paired_long <- 
  cptac_data_paired |> 
  relocate(Type) |> 
  pivot_longer(cols = c(3:ncol(cptac_data_paired)),
               names_to = "Assay", 
               values_to = "Protein_expression") |> 
  mutate(Protein_expression = as.numeric(Protein_expression)) |> 
  filter(!is.na(Protein_expression))

all_data <- 
  cancer_data_long |>
  filter(Disease %in% c("PAN", "CAD healthy"),
         Assay %in% top_30) |>
  mutate(Disease = recode(Disease, 
                          PAN = "Pancreatic cancer",
                          `CAD healthy` = "Healthy")) |> 
  select(Sample = DAid, Assay, Protein_expression = NPX, Disease) |>
  mutate(Dataset = "Disease Atlas") |> 
  bind_rows(cptac_data_paired_long |>
              filter(Assay %in% top_30) |> 
              mutate(Disease = recode(Type,
                                      Tumor = "Pancreatic cancer",
                                      Normal = "Healthy"),
                     Dataset = "CPTAC") |> 
              select(-Type)) |>
  bind_rows(ukb_data |> 
              filter(Assay %in% top_30) |> 
              left_join(ukb_meta_ext, by = "id") |> 
              select(Sample = id, Assay, Protein_expression = NPX, Disease = Group) |> 
              mutate(Dataset = "UKB",
                     Sample = as.character(Sample))) |> 
  mutate(Disease = factor(Disease, levels = c(names(pal_group), names(pal_ukb)) |> unique()))

# Disease Atlas
top_30_order <- 
  all_rank |> 
  head(30) |> 
  arrange(-logFC_healthy) |> 
  pull(Assay)


scaled_mean_top30 <- 
  # cancer_data_long |> 
  # filter(Assay %in% top_30,
  #        Disease %in% c("PAN", "CAD healthy")) |> 
  # select(DAid, Disease, Assay, NPX) |> 
  # mutate(Disease = recode(Disease,
  #                         PAN = "Pancreatic cancer",
  #                         `CAD healthy` = "Healthy")) |> 
  all_data |> 
  filter(Dataset == "Disease Atlas") |> 
  rename(NPX =Protein_expression) |> 
  group_by(Assay) |> 
  mutate(NPX = NPX-mean(NPX)/sd(NPX)) |> 
  ungroup() |> 
  group_by(Assay, Disease) |> 
  summarise(NPX = mean(NPX))
  
coord_straightpolar <- function(theta = 'x', start = 0, direction = 1, clip = "on") {
  theta <- match.arg(theta, c("x", "y"))
  r <- if (theta == "x") 
    "y"
  else "x"
  ggproto(NULL, CoordPolar, theta = theta, r = r, start = start,
          direction = sign(direction), clip = clip,
          # This is the different bit
          is_linear = function(){TRUE})
}

radar_hpa <- 
  scaled_mean_top30 %>% 
  mutate(Assay = factor(Assay, levels = top_30_order),
         Disease = factor(Disease, levels = c("Healthy", "Pancreatic cancer"))) %>%
  arrange(Assay) |> 
  ggplot(aes(Assay, NPX, color = Disease, group = Disease)) +
  coord_straightpolar(theta = 'x') +
  geom_point(size = 2, show.legend = F) +
  geom_polygon(alpha = 0.2, size = 1, fill = NA, show.legend = F) +
  #scale_y_continuous(limits = c(0.75,1)) +
  scale_color_manual(values = pal_group) +
  scale_fill_manual(values = pal_group) +
  theme_minimal() +
  xlab("") +
  ggtitle("HPA data") +
  theme(plot.title = element_text(hjust = 0.5, face = "bold"),
        axis.title.y = element_text(face = "bold"))

scaled_mean_top30_cptac <- 
  all_data |> 
  filter(Dataset == "CPTAC") |> 
  group_by(Assay) |> 
  mutate(Protein_expression = Protein_expression-mean(Protein_expression)/sd(Protein_expression)) |> 
  ungroup() |> 
  group_by(Assay, Disease) |> 
  summarise(Protein_expression = mean(Protein_expression))

radar_cptac <- 
  scaled_mean_top30_cptac %>% 
  mutate(Assay = factor(Assay, levels = top_30_order),
         Disease = factor(Disease, levels = c("Healthy", "Pancreatic cancer"))) %>%
  arrange(Assay) |> 
  ggplot(aes(Assay, Protein_expression, color = Disease, group = Disease)) +
  coord_straightpolar(theta = 'x') +
  geom_point(size = 2, show.legend = F) +
  geom_polygon(alpha = 0.2, size = 1, fill = NA, show.legend = F) +
  #scale_y_continuous(limits = c(0.75,1)) +
  scale_color_manual(values = pal_group) +
  scale_fill_manual(values = pal_group) +
  theme_minimal() +
  xlab("") +
  ggtitle("CPTAC") +
  theme(plot.title = element_text(hjust = 0.5, face = "bold"),
        axis.title.y = element_text(face = "bold"))

scaled_mean_top30_ukb <- 
  all_data |> 
  filter(Dataset == "UKB") |> 
    filter(Disease != "> 1 year after") |> 
  rename(NPX =Protein_expression) |> 
  group_by(Assay) |> 
  mutate(NPX = NPX-mean(NPX)/sd(NPX)) |> 
  ungroup() |> 
  group_by(Assay, Disease) |> 
  summarise(NPX = mean(NPX))

radar_ukb <- 
  scaled_mean_top30_ukb %>% 
  mutate(Assay = factor(Assay, levels = top_30_order),
         Disease = factor(Disease, levels = names(pal_ukb))) %>%
  arrange(Assay) |> 
  ggplot(aes(Assay, NPX, color = Disease, group = Disease)) +
  coord_straightpolar(theta = 'x') +
  geom_point(size = 2, show.legend = F) +
  geom_polygon(alpha = 0.2, size = 1, fill = NA, show.legend = F) +
  #scale_y_continuous(limits = c(0.75,1)) +
  scale_color_manual(values = pal_ukb) +
  scale_fill_manual(values = pal_ukb) +
  theme_minimal() +
  xlab("") +
  ggtitle("UKB data") +
  theme(plot.title = element_text(hjust = 0.5, face = "bold"),
        axis.title.y = element_text(face = "bold"))


radar_hpa / radar_ukb / radar_cptac
ggsave(savepath("radar_plots.pdf"), h = 13, w = 7)x

radar_hpa | radar_ukb | radar_cptac
ggsave(savepath("radar_plots_horiz.pdf"), h = 7, w = 13)
```

## 0-1 scaling

```{r}

# Disease Atlas
top_30_order <- 
  all_rank |> 
  head(30) |> 
  arrange(-logFC_healthy) |> 
  pull(Assay)


scaled_mean_top30 <- 
  # cancer_data_long |> 
  # filter(Assay %in% top_30,
  #        Disease %in% c("PAN", "CAD healthy")) |> 
  # select(DAid, Disease, Assay, NPX) |> 
  # mutate(Disease = recode(Disease,
  #                         PAN = "Pancreatic cancer",
  #                         `CAD healthy` = "Healthy")) |> 
  all_data |> 
  filter(Dataset == "Disease Atlas") |> 
  rename(NPX =Protein_expression) |> 
  group_by(Assay) |> 
  mutate(NPX = NPX/max(NPX)) |> 
  ungroup() |> 
  group_by(Assay, Disease) |> 
  summarise(NPX = mean(NPX))
  
coord_straightpolar <- function(theta = 'x', start = 0, direction = 1, clip = "on") {
  theta <- match.arg(theta, c("x", "y"))
  r <- if (theta == "x") 
    "y"
  else "x"
  ggproto(NULL, CoordPolar, theta = theta, r = r, start = start,
          direction = sign(direction), clip = clip,
          # This is the different bit
          is_linear = function(){TRUE})
}

radar_hpa <- 
  scaled_mean_top30 %>% 
  mutate(Assay = factor(Assay, levels = top_30_order),
         Disease = factor(Disease, levels = c("Healthy", "Pancreatic cancer"))) %>%
  arrange(Assay) |> 
  ggplot(aes(Assay, NPX, color = Disease, group = Disease)) +
  coord_straightpolar(theta = 'x') +
  geom_point(size = 2) +
  geom_polygon(alpha = 0.2, size = 1, fill = NA, show.legend = F) +
  #scale_y_continuous(limits = c(0.75,1)) +
  scale_color_manual(values = pal_group) +
  scale_fill_manual(values = pal_group) +
  theme_minimal() +
  xlab("") +
  ggtitle("HPA data") +
  theme(plot.title = element_text(hjust = 0.5, face = "bold"),
        axis.title.y = element_text(face = "bold"))

scaled_mean_top30_cptac <- 
  all_data |> 
  filter(Dataset == "CPTAC") |> 
  group_by(Assay) |> 
  mutate(Protein_expression = Protein_expression/max(Protein_expression)) |> 
  ungroup() |> 
  group_by(Assay, Disease) |> 
  summarise(Protein_expression = mean(Protein_expression))

radar_cptac <- 
  scaled_mean_top30_cptac %>% 
  mutate(Assay = factor(Assay, levels = top_30_order),
         Disease = factor(Disease, levels = c("Healthy", "Pancreatic cancer"))) %>%
  arrange(Assay) |> 
  ggplot(aes(Assay, Protein_expression, color = Disease, group = Disease)) +
  coord_straightpolar(theta = 'x') +
  geom_point(size = 2) +
  geom_polygon(alpha = 0.2, size = 1, fill = NA, show.legend = F) +
  #scale_y_continuous(limits = c(0.75,1)) +
  scale_color_manual(values = pal_group) +
  scale_fill_manual(values = pal_group) +
  theme_minimal() +
  xlab("") +
  ggtitle("CPTAC") +
  theme(plot.title = element_text(hjust = 0.5, face = "bold"),
        axis.title.y = element_text(face = "bold"))

scaled_mean_top30_ukb <- 
  all_data |> 
  filter(Dataset == "UKB") |> 
    filter(Disease != "> 1 year after") |> 
  rename(NPX =Protein_expression) |> 
  group_by(Disease) |> 
  mutate(NPX = NPX/max(NPX)) |> 
  ungroup() |> 
  group_by(Assay, Disease) |> 
  summarise(NPX = mean(NPX))

radar_ukb <- 
  scaled_mean_top30_ukb %>% 
  mutate(Assay = factor(Assay, levels = top_30_order),
         Disease = factor(Disease, levels = names(pal_ukb))) %>%
  arrange(Assay) |> 
  ggplot(aes(Assay, NPX, color = Disease, group = Disease)) +
  coord_straightpolar(theta = 'x') +
  geom_point(size = 2) +
  geom_polygon(alpha = 0.2, size = 1, fill = NA, show.legend = F) +
  #scale_y_continuous(limits = c(0.75,1)) +
  scale_color_manual(values = pal_ukb) +
  scale_fill_manual(values = pal_ukb) +
  theme_minimal() +
  xlab("") +
  ggtitle("UKB data") +
  theme(plot.title = element_text(hjust = 0.5, face = "bold"),
        axis.title.y = element_text(face = "bold"))


radar_hpa / radar_ukb / radar_cptac
ggsave(savepath("radar_plots.pdf"), h = 13, w = 7)

```


# PCA

```{r}
pca_data <- 
  cancer_data_wide |>
  select(DAid, Disease, top_30) |> 
  filter(Disease %in% c("PAN", "CAD healthy")) |> 
  select(-Disease) |> 
  impute_values(ID = "DAid", wide_data = T) |> 
  column_to_rownames("DAid") |> 
  scale() |> 
  t() |> 
  do_pca()

scores <- 
  pca_data@scores |> 
  as.data.frame() |> 
  rownames_to_column("DAid") |> 
  left_join(da_meta |> 
              filter(Disease %in% c("Pancreatic cancer", "CAD healthy")), 
            by = "DAid") |> 
  mutate(Disease = ifelse(Disease == "CAD healthy", "Healthy", Disease)) |> 
  ggplot(aes(PC1, PC2, color = Disease)) +
  geom_point(alpha = 0.7) +
  stat_ellipse(geom = "polygon", alpha = 0.3, aes(fill = Disease), color = NA) +
  scale_color_manual(values = pal_group) +
  scale_fill_manual(values = pal_group) +
  theme_hpa() +
  theme(legend.position = "top")

loadings <- 
  pca_data@loadings |> 
  as.data.frame() |> 
  rownames_to_column("Protein") |> 
  ggplot(aes(PC1, PC2)) +
  geom_point() +
  geom_text_repel(aes(label = Protein)) +
  theme_hpa()

scores + loadings
ggsave(savepath("pca_pancreas_top30.pdf"), h = 8, w = 10

```


## Explore proteins metadata

```{r}

pancreatic_meta


metadata_plots <- 
  function(protein) {
  stage <- 
  cancer_data_long |>
  select(DAid, Disease, Assay, NPX) |> 
  filter(Disease %in% c("PAN"),
         Assay %in% top_30) |> 
  filter(Assay == protein) |> 
  select(-Disease) |> 
  left_join(meta_pancreatic, by = "DAid") |> 
  ggplot(aes(Stage_short, NPX, color = Stage_short, fill = Stage_short)) +
  geom_quasirandom() +
  scale_fill_manual(values = pal_stage) +
  scale_color_manual(values = pal_stage) +
  geom_boxplot(alpha = 0.4, outlier.colour = NA, color = "black") +
  theme_hpa(angled = T) +
  ggtitle("Stage")

node <- 
  
cancer_data_long |>
  select(DAid, Disease, Assay, NPX) |> 
  filter(Disease %in% c("PAN"),
         Assay %in% top_30) |> 
  filter(Assay == protein) |> 
  select(-Disease) |> 
  left_join(meta_pancreatic, by = "DAid") |> 
  ggplot(aes(N, NPX, color = N, fill = N)) +
  geom_quasirandom() +
  geom_boxplot(alpha = 0.4, outlier.colour = NA, color = "black") +
  theme_hpa(angled = T) +
  ggtitle("Node")


metastasis <- 
  cancer_data_long |>
  select(DAid, Disease, Assay, NPX) |> 
  filter(Disease %in% c("PAN"),
         Assay %in% top_30) |> 
  filter(Assay == protein) |> 
  select(-Disease) |> 
  left_join(meta_pancreatic, by = "DAid") |> 
  ggplot(aes(M, NPX, color = M, fill = M)) +
  geom_quasirandom() +
  geom_boxplot(alpha = 0.4, outlier.colour = NA, color = "black") +
  theme_hpa(angled = T) +
  ggtitle("Metastasis")

sex <- 
  cancer_data_long |>
  select(DAid, Disease, Assay, NPX) |> 
  filter(Disease %in% c("PAN"),
         Assay %in% top_30) |> 
  filter(Assay == protein) |> 
  select(-Disease) |> 
  left_join(meta_pancreatic, by = "DAid") |> 
  ggplot(aes(gender, NPX, color = gender, fill = gender)) +
  geom_quasirandom() +
  geom_boxplot(alpha = 0.4, outlier.colour = NA, color = "black") +
  theme_hpa(angled = T) +
  ggtitle("Sex")

(stage | node | metastasis |sex ) +
  plot_annotation(title = protein)
  }



map(top_30, function(protein) {
  
  plot <-  metadata_plots(protein)
  ggsave(filename = savepath(paste0("metadata_plots/", protein, "_plot.png")), plot = plot, h = 4, w = 14)
  
})



metadata_plots("HGF")

```

## UKB

```{r}
complete_patients <- 
  ukb_data |>
  filter(Assay %in% top_30) |> 
  group_by(id) |> 
  summarize(n = n()) |> 
  arrange(n) |> 
  filter(n== 30)


pca_data <- 
  ukb_data |>
  filter(Assay %in% top_30,
         id %in% complete_patients$id) |> 
  select(-Disease) |> 
  pivot_wider(names_from = "Assay", values_from = "NPX") |> 
  impute_values(ID = "id", wide_data = T) |> 
  column_to_rownames("id") |> 
  scale() |> 
  t() |> 
  do_pca()

#scores <- 
  pca_data@scores |> 
  as.data.frame() |> 
  rownames_to_column("id") |> 
  left_join(ukb_meta_ext |> 
              mutate(id = as.character(id)), 
            by = "id") |> 
  ggplot(aes(PC1, PC2, color = Group)) +
  geom_point(alpha = 0.7) +
  #stat_ellipse(geom = "polygon", alpha = 0.3, aes(fill = Group), color = NA) +
  scale_color_manual(values = pal_ukb) +
  scale_fill_manual(values = pal_ukb) +
  theme_hpa() +
  theme(legend.position = "top")

loadings <- 
  pca_data@loadings |> 
  as.data.frame() |> 
  rownames_to_column("Protein") |> 
  ggplot(aes(PC1, PC2)) +
  geom_point() +
  geom_text_repel(aes(label = Protein)) +
  theme_hpa()

scores + loadings
ggsave(savepath("pca_pancreas_top30.pdf"), h = 8, w = 10)

```


# Heatmap

```{r}
ann <- 
  cancer_data_wide |> 
  filter(Disease %in% c("CAD healthy", "PAN")) |> 
  select(DAid, Disease) |> 
  column_to_rownames("DAid")

cancer_data_wide |> 
  filter(Disease %in% c("CAD healthy", "PAN")) |> 
  select(DAid, top_30) |> 
  column_to_rownames("DAid") |> 
  scale() |> 
  as.data.frame() |> 
  pheatmap(show_rownames = F, 
           clustering_method = "ward.D2",
           annotation_row = ann)

```


# Random

## Plot across cancers

```{r}
protein <- "DPY30"
long_levels <- 
  cancers_mapping |> 
  mutate(Cancer = ifelse(Cancer_code %in% c("AML", "CLL"), Cancer_code, Cancer),
         Cancer = ifelse(Cancer_code %in% c("LYMPH"), "DLBCL", Cancer),
         Cancer_code = factor(Cancer_code, levels = all_cancers_levels)) |> 
  arrange(Cancer_code) |> 
  pull(Cancer)

cancer_data_long |> 
  filter(Assay == protein,
         Disease %in% all_levels) |> 
  rename(GROUP = Disease) |> 
  mutate(Disease = case_when(GROUP == "PAN" ~ "Pancreatic cancer",
                             GROUP == "CAD healthy" ~ "Healthy",
                             T ~ "Other cancers")) |> 
  left_join(cancers_mapping |> 
              mutate(Cancer = ifelse(Cancer_code %in% c("AML", "CLL"), Cancer_code, Cancer),
                     Cancer = ifelse(Cancer_code %in% c("LYMPH"), "DLBCL", Cancer)), 
            by = c("GROUP" = "Cancer_code")) |> 
  mutate(Cancer = ifelse(GROUP == "CAD healthy", "Healthy", Cancer),
         Cancer = factor(Cancer, levels = c(long_levels, "Healthy"))) |> 
  select(DAid, Cancer, Disease, Assay, NPX) |> 
  ggplot(aes(Cancer, NPX, color = Disease, fill = Disease)) +
  geom_quasirandom(alpha = 0.4, show.legend = F) +
  geom_boxplot(alpha = 0.4, color = "grey60", outlier.color = NA) +
  scale_color_manual(values = pal_group) +
  scale_fill_manual(values = pal_group) +
  xlab("") +
  ggtitle(protein) +
  theme_hpa(angled = T)

ggsave(savepath("dpy30.png"))
```

## SSC4D

```{r}

protein <- "SSC4D"

cancer_data_long |>
  filter(Assay == "SSC4D", 
         Disease %in% c("PAN", "CAD healthy")) |>
  mutate(Disease = recode(Disease, 
                          PAN = "Pancreatic cancer",
                          `CAD healthy` = "Healthy")) |> 
  select(Sample = DAid, Protein_expression = NPX, Disease) |>
  left_join(cancer_metadata |> 
              distinct(Sample = DAid, Sex), by = "Sample") |> 
  mutate(Dataset = "Disease Atlas") |> 
  bind_rows(ukb_data |> 
              filter(Assay == protein) |> 
              left_join(ukb_meta_ext |> 
                          mutate(Sex = case_when(Sex == 0 ~ "Female",
                                                 Sex == 1 ~ "Male")), by = "id") |> 
              select(Sample = id, Protein_expression = NPX, Disease = Group, Sex) |> 
              mutate(Dataset = "UKB",
                     Sample = as.character(Sample))) |> 
  mutate(Disease = factor(Disease, levels = c(names(pal_group), names(pal_ukb)) |> unique())) |> 
  filter(!is.na(Sex)) |> 
  ggplot(aes(Disease,Protein_expression, color = Disease, fill = Disease)) +
  geom_quasirandom(alpha = 0.8) +
  geom_boxplot(alpha = 0.2, color = "black", outlier.colour = NA) +
  facet_wrap(Sex~Dataset, scales = "free") +
  theme_hpa() +
  scale_color_manual(values = c(pal_group, pal_ukb)) +
  scale_fill_manual(values = c(pal_group, pal_ukb)) +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1)) +
  xlab("") +
  ggtitle("SSC4D")

ggsave(savepath("SSC4D.png"))

```

## Volcano plot PCD

```{r}

source("scripts/functions_visualization.R")

de_df <- 
  de |> 
  select(Assay, logFC = logFC_healthy, adj.P.Val = adjpval_healthy, sig = sig_healthy)

 labels <- 
    de_df |> 
   mutate(`Pancreatic cancer database` = ifelse(Assay %in% pcd_proteins, "Yes", "No")) |> 
    top_n(n = 30, wt = -log10(adj.P.Val)) 
  
 de_df |> 
   mutate(`Pancreatic cancer database` = ifelse(Assay %in% pcd_proteins, "Yes", "No")) |> 
   ggplot(aes(x = logFC, y = -log10(adj.P.Val), color = `Pancreatic cancer database`, label = Assay)) +
   geom_point(size = 1, alpha = 0.4) + 
   geom_text_repel(data = labels, size = 2, show.legend = F) +
   geom_hline(yintercept = -log10(0.05), linetype = 'dashed', color = "darkgrey") +
   geom_vline(xintercept = -1, linetype = 'dashed', color = "darkgrey") +
   geom_vline(xintercept = 1, linetype = 'dashed', color = "darkgrey") +
   scale_color_manual(values = c("grey","red")) +
   theme_hpa() +   
   theme(axis.text = element_text(size = 8),
         legend.position = "top") 
 
 ggsave(savepath("volcano.png"))
```
